Extract question summaries into separate files
