+++ app/classifier/tasks/single/index.jsx
@@ -2,104 +2,9 @@ import { Markdown } from 'markdownz';
 import React from 'react';
 import GenericTask from '../generic';
 import GenericTaskEditor from '../generic-editor';
+import SingleChoiceSummary from './summary';
 
 const NOOP = Function.prototype;
-
-class Summary extends React.Component {
-  constructor(props) {
-    super(props);
-    this.expand = this.expand.bind(this);
-    this.collapse = this.collapse.bind(this);
-    this.state = {
-      expanded: this.props.expanded
-    };
-  }
-
-  expand() {
-    this.setState({ expanded: true });
-  }
-
-  collapse() {
-    this.setState({ expanded: false });
-  }
-
-  render() {
-    const { annotation, task, translation } = this.props;
-    let toggleButton = <button type="button" className="toggle-more" onClick={this.expand}>More</button>;
-    let answers = <div className="answer">No answer</div>;
-    if (!this.state.expanded && annotation.value !== null) {
-      answers = (
-        <div className="answer">
-          <i className="fa fa-check-circle-o fa-fw"></i>
-          <Markdown tag="span" inline={true}>{translation.answers[annotation.value].label}</Markdown>
-        </div>
-      );
-    }
-    if (this.state.expanded) {
-      toggleButton = <button type="button" className="toggle-more" onClick={this.collapse}>Less</button>;
-      answers = [];
-      for (const [i, answer] of task.answers.entries()) {
-        if (!answer._key) {
-          answer._key = Math.random();
-        }
-        let icon = <i className="fa fa-circle-o fa-fw"></i>;
-        if (i === annotation.value) {
-          icon = <i className="fa fa-check-circle-o fa-fw"></i>;
-        }
-        answers.push(
-          <div key={answer._key} className="answer">
-            {icon}
-            <Markdown tag="span" inline={true}>{translation.answers[i].label}</Markdown>
-          </div>
-        );
-      }
-    }
-    return (
-      <div>
-        <div className="question">
-          <Markdown>
-            {this.props.task.question}
-          </Markdown>
-          {toggleButton}
-        </div>
-        <div className="answers">
-          {answers}
-        </div>
-      </div>
-    );
-  }
-}
-
-Summary.propTypes = {
-  task: React.PropTypes.shape(
-    {
-      answers: React.PropTypes.array,
-      question: React.PropTypes.string
-    }
-  ),
-  translation: React.PropTypes.shape({
-    characteristics: React.PropTypes.object,
-    choices: React.PropTypes.object,
-    questions: React.PropTypes.object
-  }).isRequired,
-  annotation: React.PropTypes.shape(
-    { value: React.PropTypes.number }
-  ).isRequired,
-  expanded: React.PropTypes.bool
-};
-
-Summary.defaultProps = {
-  task: {
-    answers: [],
-    question: ''
-  },
-  translation: {
-    answers: [],
-    question: '',
-    help: ''
-  },
-  expanded: false
-};
 
 export default class SingleChoiceTask extends React.Component {
   constructor(props) {
@@ -156,7 +61,7 @@ export default class SingleChoiceTask extends React.Component {
 
 // Define the static methods and values
 SingleChoiceTask.Editor = GenericTaskEditor;
-SingleChoiceTask.Summary = Summary;
+SingleChoiceTask.Summary = SingleChoiceSummary;
 SingleChoiceTask.getDefaultTask = () => {
   return {
     type: 'single',
